+++ frontend/src/pages/products/CatalogPage.jsx
@@ -1,5 +1,6 @@
 import { useEffect } from "react"
 import ProductCatalog from "./ProductCatalog"
+import { Container } from "react-bootstrap"
 
 const CatalogPage = () => {
   // Scroll to the top of the page when the component mounts
@@ -9,9 +10,9 @@ const CatalogPage = () => {
 
   // Render the Catalog component
   return (
-    <div className="d-flex flex-column" style={{ minHeight: "100vh" }}>
+    <Container fluid className="d-flex flex-column min-vh-100 p-0">
       <ProductCatalog showDetails={true} />
-    </div>
+    </Container>
   )
 }