+++ frontend/src/utils/axiosInstance.js
@@ -1,7 +1,10 @@
 import axios from "axios"
 
 const axiosInstance = axios.create({
-  baseURL: "http://localhost:5000/api",
+  // Dev
+  // baseURL: "http://localhost:5000/api",
+  // Prod
+  baseURL: "https://fura-solutions-backend.onrender.com/api",
 })
 
 export default axiosInstance