+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [v0.2.1] - 2025-02-16
+
+### Added
+- Deployment of backend on the cloud and integration with the github page
+
+### Fixed
+- Error when trying to add products to cart
+- Login error due to localhost address when deployed
+
+### Changed
+- All page headers have same margins
+
+### Removed
+- Unnecessary logs
+- Flag icons package using pngs instead
+
 ## [v0.2.0] - 2025-02-03
 
 ### Added